pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "maxun-python-sdk"
version = "0.0.1"
description = "Python SDK for Maxun - web automation and data extraction"
requires-python = ">=3.8"
dependencies = ["httpx>=0.24.0", "python-dotenv>=1.0.0"]

[project.optional-dependencies]
anthropic = ["anthropic>=0.18.0"]
openai = ["openai>=1.0.0"]
all = ["anthropic>=0.18.0", "openai>=1.0.0"]

[tool.setuptools]
packages = ["maxun", "maxun.builders", "maxun.llms", "maxun.llms.providers"]

[tool.setuptools.package-dir]
maxun = "."
